HPCC-13488 ps -ef will reveal passwords
Add prompt for password if only the username presented in command line
