+++ dali/dfuplus/dfuplus.cpp
@@ -25,7 +25,9 @@
 #include "bindutil.hpp"
 #include "ws_fs.hpp"
 #include "ws_dfu.hpp"
-
+#if defined( __linux__) || defined(__FreeBSD__)
+#include "termios.h"
+#endif
 #define WAIT_SECONDS 30
 
 #ifdef DAFILESRV_LOCAL
@@ -87,6 +89,69 @@ public:
     unsigned idleTime() { return server.get()?server->idleTime():0; }
 
 };
+
+void CDfuPlusHelper::promptFor(const char *prompt, const char *prop, bool hide, IProperties * globals)
+{
+    StringBuffer result;
+    fprintf(stdout, "%s", prompt);
+    fflush(stdout);
+    if (hide)
+    {
+#ifdef _WIN32
+        HANDLE hStdIn = GetStdHandle(STD_INPUT_HANDLE);
+        DWORD dwInputMode;
+        GetConsoleMode(hStdIn, &dwInputMode);
+        SetConsoleMode(hStdIn, dwInputMode & ~ENABLE_LINE_INPUT & ~ENABLE_ECHO_INPUT);
+        loop
+        {
+            /* read a character from the console input */
+            char ch;
+            DWORD dwRead;
+            if (!ReadFile(hStdIn, &ch, sizeof(ch), &dwRead, NULL))
+                break;
+            if (ch == '\n' || ch=='\r' || !ch)
+                break;
+            result.append(ch);
+        }
+        SetConsoleMode(hStdIn, dwInputMode);
+#else
+        int fn = fileno(stdin);
+#ifdef __linux__
+        struct termio t;
+        /* If ioctl fails, we're probably not connected to a terminal. */
+        if(!ioctl(fn, TCGETA, &t))
+        {
+            t.c_lflag &= ~ECHO;
+            ioctl(fn, TCSETA, &t);
+        }
+#endif
+        loop
+        {
+            char ch = fgetc(stdin);
+            if (ch == '\n' || ch=='\r' || !ch)
+                break;
+            result.append(ch);
+        }
+#ifdef __linux__
+        if(!ioctl(fn, TCGETA, &t))
+        {
+            t.c_lflag |= ECHO;
+            ioctl(fn, TCSETA, &t);
+        }
+#endif
+#endif
+        printf("\n");
+    }
+    else
+    {
+        char buf[100];
+        if (fgets(buf, 100, stdin))
+            result.append(buf);
+        if (result.length() && result.charAt(result.length()-1)=='\n')
+            result.remove(result.length()-1, 1);
+    }
+    globals->setProp(prop, result);
+}
 
 bool CDfuPlusHelper::runLocalDaFileSvr(SocketEndpoint &listenep,bool requireauthenticate, unsigned timeout)
 {
@@ -178,6 +243,13 @@ CDfuPlusHelper::CDfuPlusHelper(IProperties* _globals,   CDfuPlusMessagerIntercep
 
     const char* username = globals->queryProp("username");
     const char* password = globals->queryProp("password");
+    if ( username && *username && (!password || !*password))
+    {
+        VStringBuffer prompt("%s's password: ", username);
+        promptFor(prompt, "password", true, globals);
+        password = globals->queryProp("password");
+    }
+
     sprayclient->setUsernameToken(username, password, NULL);
     dfuclient->setUsernameToken(username, password, NULL);
